changed timestamp type int -> bigint

+++ database/migrations/20181025023225_create_reminders_table.ts
@@ -7,7 +7,7 @@ exports.up = (knex: Knex) => {
       .integer('chatid')
       .unique()
       .notNullable();
-    table.integer('timestamp').notNullable();
+    table.bigInteger('timestamp').notNullable();
     table.string('text').notNullable();
     table.string('asker').notNullable();
   });
